Fix test unit.
